Cleaning up SmashGG code.
